# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Service for Codex",
        "title": "API Service",
        "contact": {}
    },
    "host": "localhost",
    "basePath": "/api/v1",
    "paths": {
...
        "/private/user/connect": {
            "post": {
                "description": "Connects Wallet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Connect Wallet To User",
                "parameters": [
                    {
                        "description": "New User Wallet",
                        "name": "newWallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserAuthWallet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
        "/private/user/profile": {
            "get": {
                "description": "Retrieves users profile.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Updates users profile.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Profile",
                "parameters": [
                    {
                        "description": "New User Profile",
                        "name": "newUserProfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserProfileUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.UserAuthWallet": {
            "type": "object",
            "required": [
                "message",
                "publicKeyBase58",
                "signatureBase58"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "publicKeyBase58": {
                    "type": "string"
                },
                "signatureBase58": {
                    "type": "string"
                }
            }
        },
...
        "dto.UserProfileUpdateDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 30
                },
                "surname": {
                    "type": "string",
                    "maxLength": 30
                }
            }
        },
...
        "response.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "dataCount": {
                    "type": "integer"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}